moksh-orchestrator/pyproject.toml
[tool.poetry]
name = "moksh-orchestrator"
version = "0.1.0"
description = ""
authors = ["gmalho <gaurav.malhotra@nike.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
yamale = "^2.1.0"
"ruamel.yaml" = "^0.16.10"
mypy = "^0.780"
pinject = "^0.14.1"
python-interface = "^1.6.0"
dpath = "^2.0.1"
metaflow = "^2.0.5"
Cheetah3 = "^3.2.5"
dask = "^2.18.1"



[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
